Moved check_help_page_urls to audit_checks and edited imports
